requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
boto3>=1.34.0
rasterio>=1.3.0
pyshp>=2.3.0
shapely>=2.0.0
pyproj>=3.6.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.8.0
streamlit>=1.28.0
streamlit-image-coordinates>=0.1.9
streamlit-drawable-canvas>=0.9.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
scikit-learn>=1.3.0
